fiapx-api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9
redis>=5.0.1
pydantic[email]>=2.5.3
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.2.1
python-multipart>=0.0.6
boto3>=1.34.14
structlog>=24.1.0
httpx>=0.26.0
prometheus-client>=0.19.0
pika>=1.3.2
